+++ .vscode/settings.json
@@ -49,6 +49,7 @@
     "AFDR",
     "affero",
     "agpl",
+    "allclose",
     "Aloney",
     "analysis",
     "angiographic",
@@ -267,6 +268,7 @@
     "Kerns",
     "keyrings",
     "krekel",
+    "largebinary",
     "libjpeg",
     "Linac",
     "linac's",
