+++ lib/pymedphys/_mosaiq/mock/__init__.py
@@ -1,0 +1,1 @@
+"""A module for mocking a Mosaiq database"""
