+++ lib/pymedphys/_mosaiq/mock/data/types_map.toml
