+++ lib/pymedphys/_mosaiq/mock/extract_from_bak.py
@@ -1,7 +1,7 @@
-"""Fleshes out CSV files found within this directory from a .bak file.
+"""Fleshes out CSV files found within the data directory from a .bak file.
 
 By design, only tables and columns that have been defined within
-`types_map.toml` are extracted so as to not expose any more of the
+`data/types_map.toml` are extracted so as to not expose any more of the
 underlying database schema than what is already within the repository.
 """
