+++ lib/pymedphys/tests/mosaiq/test_can_we_db.py
@@ -15,16 +15,16 @@
 
 from pymedphys._imports import pymssql, pytest
 
-from . import _connect
+from pymedphys._mosaiq.mock import connect
 
 
 @pytest.mark.mosaiqdb
 def test_can_we_db():
     conn = pymssql.connect(
-        _connect.MSQ_SERVER,
-        port=_connect.MSQ_PORT,
-        user=_connect.SA_USER,
-        password=_connect.SA_PASSWORD,
+        connect.MSQ_SERVER,
+        port=connect.MSQ_PORT,
+        user=connect.SA_USER,
+        password=connect.SA_PASSWORD,
     )
     cursor = conn.cursor()
     cursor.execute("select * from sys.databases")
